Setup Programmer project.
The defines for __CMSIS__, __CODE_RED__ and __REDLIB__
are removed from the Programmer project. The advantage
of this is that the startup file will now not include
a call to systemInit() in the CMSIS and not include
the code read redlib library.

The compiler now searches for include files in
Program/inc.