Add bootloader loops in bootloading mode if no user application available.
The bootloader checks if there is a user application available.
If the timer has run out, and the bootloader has not been set to bootloading
mode, the bootloader checks if there is a user application available.
If not, the bootloader keeps looping and remains available to messages of
the programmer.