Remove bug in uart driver.
Some piece of code that checked whether the UART receive buffer was full,
was not correct. It fired when not the total buffer capacity was used,
but the total capacity minus one. This caused a wrap around of data in the buffer,
which becomes a problem when receiving 4kB blocks.